Cleaning up liquibase.util tests and code - Removed now-unused Main help.xml and supporting classes - Moved PropertyUtil to ObjectUtil - Renamed StringUtil.splitByCharacterType() to splitCamelCase()
