+++ liquibase-core/src/main/java/liquibase/util/ObjectUtil.java
@@ -5,6 +5,10 @@ import liquibase.statement.DatabaseFunction;
 import liquibase.statement.SequenceCurrentValueFunction;
 import liquibase.statement.SequenceNextValueFunction;
 
+import java.beans.IntrospectionException;
+import java.beans.Introspector;
+import java.beans.PropertyDescriptor;
+import java.beans.PropertyVetoException;
 import java.io.File;
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
@@ -19,6 +23,8 @@ import java.util.*;
  */
 public class ObjectUtil {
 
+    private static final List<BeanIntrospector> introspectors = new ArrayList<>(Arrays.asList(new DefaultBeanIntrospector(), new FluentPropertyBeanIntrospector()));
+    
     /**
      * Cache for the methods of classes that we have been queried about so far.
      */
@@ -393,4 +399,100 @@ public class ObjectUtil {
             return value;
         }
     }
+
+    public static PropertyDescriptor[] getDescriptors(Class<?> targetClass) throws IntrospectionException {
+        IntrospectionContext context = new IntrospectionContext(targetClass);
+        for (BeanIntrospector introspector : introspectors) {
+            introspector.introspect(context);
+        }
+        return context.getDescriptors();
+    }
+
+
+    private interface BeanIntrospector {
+        void introspect(IntrospectionContext context) throws IntrospectionException;
+    }
+
+    private static class DefaultBeanIntrospector implements BeanIntrospector {
+        @Override
+        public void introspect(IntrospectionContext context) throws IntrospectionException {
+            PropertyDescriptor[] descriptors = Introspector.getBeanInfo(context.getTargetClass()).getPropertyDescriptors();
+            if (descriptors != null) {
+                context.addDescriptors(descriptors);
+            }
+        }
+    }
+
+    private static class FluentPropertyBeanIntrospector implements BeanIntrospector {
+        @Override
+        public void introspect(IntrospectionContext context) throws IntrospectionException {
+            for (Method method : context.getTargetClass().getMethods()) {
+                try {
+                    Class<?>[] argTypes = method.getParameterTypes();
+                    int argCount = argTypes.length;
+                    if ((argCount == 1) && method.getName().startsWith("set")) {
+                        String propertyName = Introspector.decapitalize(method.getName().substring(3));
+                        if (!"class".equals(propertyName)) {
+                            PropertyDescriptor pd = context.getDescriptor(propertyName);
+                            boolean setWriteMethod = false;
+                            if (pd == null) {
+                                pd = new PropertyDescriptor(propertyName, null, method);
+                                context.addDescriptor(pd);
+                                setWriteMethod = true;
+                            } else if ((pd.getWriteMethod() == null) && (pd.getReadMethod() != null) && (pd.getReadMethod
+                                    ().getReturnType() == argTypes[0])) {
+
+                                pd.setWriteMethod(method);
+                                setWriteMethod = true;
+                            }
+                            if (setWriteMethod) {
+                                for (Class<?> type : method.getExceptionTypes()) {
+                                    if (type == PropertyVetoException.class) {
+                                        pd.setConstrained(true);
+                                        break;
+                                    }
+                                }
+                            }
+                        }
+                    }
+                } catch (IntrospectionException ignored) {
+                }
+            }
+        }
+    }
+
+    private static class IntrospectionContext {
+        private final Class<?> targetClass;
+        private final Map<String, PropertyDescriptor> descriptors = new HashMap<>();
+
+        public IntrospectionContext(Class<?> targetClass) {
+            if (targetClass == null) {
+                throw new NullPointerException();
+            }
+            this.targetClass = targetClass;
+        }
+
+        public void addDescriptor(PropertyDescriptor descriptor) {
+            descriptors.put(descriptor.getName(), descriptor);
+        }
+
+        public void addDescriptors(PropertyDescriptor[] descriptors) {
+            for (PropertyDescriptor descriptor : descriptors) {
+                addDescriptor(descriptor);
+            }
+        }
+
+        public PropertyDescriptor getDescriptor(String name) {
+            return descriptors.get(name);
+        }
+
+        public PropertyDescriptor[] getDescriptors() {
+            return descriptors.values().toArray(new PropertyDescriptor[descriptors.values().size()]);
+        }
+
+        public Class<?> getTargetClass() {
+            return targetClass;
+        }
+    }
+
 }
