--- liquibase-core/src/main/java/liquibase/util/beans/BeanIntrospector.java
@@ -1,7 +1,0 @@
-package liquibase.util.beans;
-
-import java.beans.IntrospectionException;
-
-public interface BeanIntrospector {
-    void introspect(IntrospectionContext context) throws IntrospectionException;
-}
